+++ packages/layout/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.34.1](https://github.com/ant-design/pro-components/compare/@ant-design/pro-layout@6.34.0...@ant-design/pro-layout@6.34.1) (2022-03-08)
+
+**Note:** Version bump only for package @ant-design/pro-layout
+
 # [6.34.0](https://github.com/ant-design/pro-components/compare/@ant-design/pro-layout@6.33.0...@ant-design/pro-layout@6.34.0) (2022-03-04)
 
 ### Bug Fixes