+++ packages/table/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.66.4](https://github.com/ant-design/pro-components/compare/@ant-design/pro-table@2.66.3...@ant-design/pro-table@2.66.4) (2022-03-08)
+
+### Bug Fixes
+
+- **table:** fix no fieldProps injected error ([#4787](https://github.com/ant-design/pro-components/issues/4787)) ([c4e835d](https://github.com/ant-design/pro-components/commit/c4e835dd60e56a5ca0ee2399f03c7c0c1c63e971))
+
+## [2.66.3](https://github.com/ant-design/pro-components/compare/@ant-design/pro-table@2.66.2...@ant-design/pro-table@2.66.3) (2022-03-07)
+
+### Bug Fixes
+
+- **table:** default close table auto focus first input ([a1fe1cb](https://github.com/ant-design/pro-components/commit/a1fe1cb73fbc5e545484acceac7733c0cc0cad57))
+
+## [2.66.2](https://github.com/ant-design/pro-components/compare/@ant-design/pro-table@2.66.1...@ant-design/pro-table@2.66.2) (2022-03-07)
+
+**Note:** Version bump only for package @ant-design/pro-table
+
 ## [2.66.1](https://github.com/ant-design/pro-components/compare/@ant-design/pro-table@2.66.0...@ant-design/pro-table@2.66.1) (2022-03-04)
 
 **Note:** Version bump only for package @ant-design/pro-table